crowdsourcing_requirements.txt
# This pip requirements file contains all the basic requirements of
# django-crowdsourcing. To run this file you will need to
# $ easy_install pip
# $ pip install -r crowdsourcing_requirements.txt

# Python Image Library for processing image answers.
#PIL>=1.1.6
pillow

# for OEmbed
python-oembed

# for image thumbnailing
sorl-thumbnail==3.2.5

# for ordering inlines
django-positions
#-e git://github.com/jpwatts/django-positions.git#egg=django-positions

# for unicode csv writer support
unicodecsv

# For geocoding addresses and address lookups
geopy

# For building the documentation.
# Sphinx

# For Flickr support. Crowdsourcing is fairly graceful if you don't install
# flickrapi.
flickrapi

# For asynchronously synching Flickr. Sync synchronously by default.
# carrot>=0.10.3
# celery==1.0.2

# templating
django_forms_bootstrap
